- Added Some Asserts
- Created Main Menu
- Fixed Back Teleport Glitch
- Fixed Position of Player Teleport to Other Worlds
