+++ Non-Euclidean Test/Assets/MainMenuAssets/VideoTexture.renderTexture.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dad9d90df3e2256438dca7effa4408c6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
