+++ Non-Euclidean Test/Assets/Mat/Screen.mat
@@ -57,6 +57,6 @@ Material:
     - _Strength: 200
     m_Colors:
     - _Offset: {r: 0, g: 0, b: 0, a: 0}
-    - _Tiling: {r: 2, g: 2, b: 0, a: 0}
+    - _Tiling: {r: 1, g: 1, b: 0, a: 0}
     - _Tint_Color: {r: 0.38425595, g: 0.4513256, b: 0.5188679, a: 0}
   m_BuildTextureStacks: []
