+++ Non-Euclidean Test/Assets/Scenes/Start.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2f89d10a3bb7fe04eb34a0c0a4d6b80a
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
