+++ Non-Euclidean Test/Assets/Script/Tele.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ac9175cb07a661045bfaf45525addb5b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
