+++ Non-Euclidean Test/Assets/Script/Tele/TeleCollideStop.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4ce0899f7443f224ea4ec968712469e6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
