+++ Non-Euclidean Test/Assets/Script/Tele/TeleRT.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+
+public class TeleRT : MonoBehaviour
+{
+    public Camera CamToOne;
+    public Material matOne;
+
+    private void Start()
+    {
+        if (CamToOne.targetTexture != null)
+        {
+            CamToOne.targetTexture.Release();
+        }
+        CamToOne.targetTexture = new RenderTexture(Screen.width, Screen.height, 24);
+        matOne.mainTexture = CamToOne.targetTexture;
+    }
+}
