--- Non-Euclidean Test/Assets/Script/TeleGate.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4c018279a75d9064d94096039ba9dcd6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
