--- Non-Euclidean Test/Assets/Script/TeleGateWCam.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dd325a4aec378f443a9b2c888de2d1a0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
