--- Non-Euclidean Test/Assets/Script/TeleRT.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8aaf723ee0df92544ad15691ea123e08
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
