+++ Non-Euclidean Test/Assets/Shader/Screen.shadergraph
@@ -1481,10 +1481,10 @@
         "m_Expanded": true,
         "m_Position": {
             "serializedVersion": "2",
-            "x": -1071.0,
-            "y": 168.0,
-            "width": 207.99993896484376,
-            "height": 374.00006103515627
+            "x": -1071.3333740234375,
+            "y": 168.6666717529297,
+            "width": 209.333251953125,
+            "height": 375.99993896484377
         }
     },
     "m_Slots": [
